requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.7.0
astroid==2.6.6
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dateparser==1.0.0
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.1.0
numpy==1.21.1
pycodestyle==2.7.0
pylint==2.9.6
python-binance==1.0.10
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
toml==0.10.2
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.6
websockets==9.1
Werkzeug==2.0.1
wrapt==1.12.1
yarl==1.7.2
